refactor(pom): line / arrow の直線描画を addStraightLine helper に共通化
端点座標から pptxgenjs line shape (左上座標 + flipH/flipV) への変換と
line スタイルの組み立てが line.ts / arrow.ts でほぼ同一だったため、
utils/straightLine.ts に集約する。resolveArrowType も同 helper に移動。
変換は同一のため出力挙動は変わらない。

Refs #809
